portable verifier: one Rust core behind a dependency-free C ABI (wasm32 + native)
The verifier now runs where readers are, without rewriting it. `crates/portable_verifier`
exposes `fb_init` / `fb_verify` / `fb_alloc` / `fb_free` as raw `extern "C"` (no wasm-bindgen,
no new dependencies), so the same core serves WebAssembly on the web and Dart FFI on mobile
while the CLI is unchanged.

The verification mathematics is untouched: verify_execution, WHIR, the sub-protocols and the
tables have no diff. What changed is infrastructure that has no wasm32 equivalent --
`backend/parallel` gains a real serial dispatcher (the pre-existing `forbid-parallelism`
feature was NOT a fallback: it bans dispatch and asserts, so it was kept as a guard and a
serial implementation added), `zk-alloc` gains a plain-heap path beside the mmap arena, and
system-info answers on targets without sysctl.

- `fb_init(bytes)` rebuilds the bytecode from the artifact, recomputes its hash and rejects
  anything that does not equal a compile-time pinned VK hash; dimensions and trailing bytes
  are checked first. A wrong or hostile artifact cannot influence verification.
- `fb_verify` bounds every input, requires envelope v3, binds the caller's row count and blob
  id, and reconstructs rows with the signature field zeroed -- a verifier never needs one.
- `examples/strip-verifier-cache` clears hints and debug maps from a compiled cache, keeps
  every instruction, and re-checks the hash: 23.93 MB -> 16.07 MB raw, 2.91 -> 2.13 MB gzipped.
  The result is verification-only and must not be given to a prover.

Measured (arm64 macOS; golden 16-row, one-leaf envelope, 183,611 bytes):

| | wasm32 (Bun / Node) | native |
|---|---|---|
| module | 1,149,309 B raw, 306,613 B gzip-9 | |
| verify | 205 / 201 ms median | 141 ms |
| init | 5.8 / 9.8 s | 4.0 s |
| linear memory | 274 MiB full cache, 214.75 MiB stripped | 341 MiB peak RSS |

Accepts the committed golden envelope and rejects changed rows, truncation and a wrong VK, in
both runtimes and in a host dylib and static binary. Startup, not download, is the cost: it is
dominated by rebuilding and hashing the 2^24-cell instruction table, whose 64 MiB in-memory
floor a verifier-only VK representation could cut. That is mapped, not implemented.

+++ crates/xmss/src/lib.rs
@@ -3,7 +3,9 @@ use backend::{DIGEST_LEN_FE, KoalaBear, POSEIDON1_WIDTH, PrimeCharacteristicRing
 
 #[cfg(any(test, feature = "test-utils"))]
 pub mod signers_cache;
+#[cfg(feature = "ssz")]
 mod ssz_serialization;
+#[cfg(feature = "ssz")]
 pub use ssz_serialization::{PUB_KEY_SSZ_LEN, SIGNATURE_SSZ_LEN};
 mod wots;
 // The rest of the WOTS layer (one-time secret keys, chain walking, encoding) is a private
@@ -111,7 +113,7 @@ pub(crate) fn poseidon_prf(domain: u32, seed: &[u8; 32], indices: [usize; 2]) ->
     input[9] = F::from_usize(high_bits);
 
     for (i, &idx) in indices.iter().enumerate() {
-        assert!(idx < 1 << 60);
+        assert!((idx as u64) < 1u64 << 60);
         input[10 + 2 * i] = F::from_usize(idx & mask);
         input[11 + 2 * i] = F::from_usize(idx >> 30);
     }
